test(platform-api): split workflow/agent mutation group tests into per-handler siblings
api-boundaries gate requires every handler file to have a sibling
__tests__/<name>.test.ts. Split the group test files (agents-mutations,
workflows-mutations, bulk-runs) into per-handler test files using the
existing it() blocks verbatim. Add tests for previously uncovered
handlers (oauth-tokens, copy-workflow, get-run-count, _run-active).

Fix a strict-schema bug in sanitize() exposed by the new oauth-tokens
test: PublicAgentOAuthTokenSchema is .strict(), so passing entries with
accessToken/serverName threw 'Unrecognized keys' on .parse().
