+++ packages/platform-api/src/handlers/agents/__tests__/create-agent.test.ts
@@ -1,0 +1,53 @@
+import { describe, it, expect, beforeEach } from 'vitest';
+import {
+  InMemoryAgentDefinitionRepository,
+  InMemoryAuditRepository,
+  InMemoryProcessInstanceRepository,
+  resetFactorySequence,
+} from '@mediforce/platform-core/testing';
+import { createAgent } from '../create-agent.js';
+import {
+  createTestScope,
+  userCaller,
+} from '../../../repositories/__tests__/create-test-scope.js';
+
+describe('createAgent handler', () => {
+  let agentDefinitionRepo: InMemoryAgentDefinitionRepository;
+  let auditRepo: InMemoryAuditRepository;
+
+  beforeEach(() => {
+    resetFactorySequence();
+    agentDefinitionRepo = new InMemoryAgentDefinitionRepository();
+    const instanceRepo = new InMemoryProcessInstanceRepository();
+    auditRepo = new InMemoryAuditRepository(instanceRepo);
+  });
+
+  function buildScope(namespaces = ['team-alpha']) {
+    return createTestScope({
+      agentDefinitionRepo,
+      auditRepo,
+      caller: userCaller('u-1', namespaces),
+    });
+  }
+
+  it('createAgent stores the agent + emits audit', async () => {
+    const scope = buildScope();
+    const { agent } = await createAgent(
+      {
+        kind: 'plugin',
+        name: 'Bob',
+        iconName: 'Bot',
+        description: 'd',
+        foundationModel: 'm',
+        systemPrompt: 'p',
+        inputDescription: 'in',
+        outputDescription: 'out',
+        skillFileNames: [],
+        namespace: 'team-alpha',
+        visibility: 'private',
+      },
+      scope,
+    );
+    expect(agent.name).toBe('Bob');
+  });
+});
