+++ packages/platform-api/src/handlers/processes/__tests__/bulk-cancel-runs.test.ts
@@ -6,7 +6,6 @@ import {
   resetFactorySequence,
 } from '@mediforce/platform-core/testing';
 import { bulkCancelRuns } from '../bulk-cancel-runs.js';
-import { bulkArchiveRuns } from '../bulk-archive-runs.js';
 import { createTestScope, userCaller } from '../../../repositories/__tests__/create-test-scope.js';
 
 describe('bulkCancelRuns handler', () => {
@@ -75,38 +74,5 @@ describe('bulkCancelRuns handler', () => {
     const { results } = await bulkCancelRuns({ runIds: ['a'] }, scope);
 
     expect(results[0]).toMatchObject({ id: 'a', status: 'error' });
-  });
-});
-
-describe('bulkArchiveRuns handler', () => {
-  let instanceRepo: InMemoryProcessInstanceRepository;
-  let auditRepo: InMemoryAuditRepository;
-
-  beforeEach(() => {
-    resetFactorySequence();
-    instanceRepo = new InMemoryProcessInstanceRepository();
-    auditRepo = new InMemoryAuditRepository(instanceRepo);
-  });
-
-  it('archives each eligible run and reports errors for active runs', async () => {
-    await instanceRepo.create(
-      buildProcessInstance({ id: 'a', namespace: 'team-alpha', status: 'completed' }),
-    );
-    await instanceRepo.create(
-      buildProcessInstance({ id: 'b', namespace: 'team-alpha', status: 'running' }),
-    );
-    const scope = createTestScope({
-      instanceRepo,
-      auditRepo,
-      caller: userCaller('u-1', ['team-alpha']),
-    });
-
-    const { results } = await bulkArchiveRuns({ runIds: ['a', 'b'] }, scope);
-
-    expect(results[0]).toEqual({ id: 'a', status: 'ok' });
-    expect(results[1]).toMatchObject({ id: 'b', status: 'error' });
-
-    const eventsA = await auditRepo.getByProcess('a');
-    expect(eventsA.map((e) => e.action)).toEqual(['instance.archived']);
   });
 });
